+++ src/AppendStream.php
@@ -1,6 +1,8 @@
 <?php
 namespace GuzzleHttp\Stream;
 
+use GuzzleHttp\Stream\Exception\CannotAttachException;
+
 /**
  * Reads from multiple streams, one after the other.
  *
@@ -14,6 +16,7 @@ class AppendStream implements StreamInterface
     private $seekable = true;
     private $current = 0;
     private $pos = 0;
+    private $detached = false;
 
     /**
      * @param StreamInterface[] $streams Streams to decorate. Each stream must
@@ -57,9 +60,9 @@ class AppendStream implements StreamInterface
         $this->streams[] = $stream;
     }
 
-    public function getContents($maxLength = -1)
+    public function getContents()
     {
-        return Utils::copyToString($this, $maxLength);
+        return Utils::copyToString($this);
     }
 
     /**
@@ -86,6 +89,12 @@ class AppendStream implements StreamInterface
     public function detach()
     {
         $this->close();
+        $this->detached = true;
+    }
+
+    public function attach($stream)
+    {
+        throw new CannotAttachException();
     }
 
     public function tell()
@@ -204,8 +213,8 @@ class AppendStream implements StreamInterface
         return false;
     }
 
-    public function flush()
+    public function getMetadata($key = null)
     {
-        return false;
+        return $key ? null : [];
     }
 }