+++ src/CachingStream.php
@@ -1,12 +1,13 @@
 <?php
 namespace GuzzleHttp\Stream;
+
 use GuzzleHttp\Stream\Exception\SeekException;
 
 /**
  * Stream decorator that can cache previously read bytes from a sequentially
  * read stream.
  */
-class CachingStream implements StreamInterface, MetadataStreamInterface
+class CachingStream implements StreamInterface
 {
     use StreamDecoratorTrait;